Makes Bank Accounts Accessible (While Character is not-ingame)
The idea of this is so that bank accounts can be accessed when people aren't around by the game's code.

Which means this allows devs to make code where people recieve or lose money while offline. This might mean we'll have to put in some temporary transisition code, but it should be fun and good once we get it working.